local

+++ DnD_Map_Maker_3D/Assets/Scripts/Login.cs
@@ -2,6 +2,7 @@ using System;
 using System.Diagnostics;
 using System.IO;
 using System.Net;
+using System.Threading;
 using TMPro;
 using UnityEngine;
 using UnityEngine.SceneManagement;
@@ -12,14 +13,12 @@ using Debug = UnityEngine.Debug;
 /// </summary>
 public class Login : MonoBehaviour
 {
-    [SerializeField]
-    private TMP_InputField serverIp;
+    [SerializeField] private TMP_InputField serverIp;
 
-    [SerializeField]
-    private PopupController popup;
+    [SerializeField] private PopupController popup;
 
-    [SerializeField]
-    private TMP_InputField username;
+    [SerializeField] private TMP_InputField username;
+
     /// <summary>
     /// Gets called when the user clicks the connect button.
     /// Test the server connection and if it works, loads the next scene
@@ -28,20 +27,32 @@ public class Login : MonoBehaviour
     {
         try
         {
+            Debug.Log("http://" + serverIp.text + ":443/DnD/TestConnection");
             var webRequest = WebRequest.Create("http://" + serverIp.text + ":443/DnD/TestConnection");
             webRequest.Proxy = null;
-            var responseString =  Connect(webRequest); 
+            var responseString = Connect(webRequest);
+            Debug.Log(responseString);
             if ("Connection erstellt" == responseString)
             {
                 DataContainer.ServerIP = serverIp.text;
-                DataContainer.WebserviceConnection = Process.Start(@"..\Int5.DnD3D.WebClient\Int5.DnD3D.WebClient\bin\Debug\net6.0\Int5.DnD3D.WebClient.exe", DataContainer.ServerIP + " " + username.text);
-                DataContainer.WebserviceConnection!.StartInfo.CreateNoWindow = true;
-                while (!File.Exists(Path.GetTempPath() + @$"/DnD/0"))
-                { }
-                string content = File.ReadAllText(Path.GetTempPath() + "/DnD/" + 0);
-                DataContainer.ClientId = Guid.Parse(content.Substring(1, content.Length - 1));
-                File.Delete(Path.GetTempPath() + "/DnD/0");
-                SceneManager.LoadScene("SampleScene");
+                string arg = DataContainer.ServerIP + " " + username.text;
+                DataContainer.WebserviceConnection.StartInfo.Arguments = arg;
+                DataContainer.WebserviceConnection.StartInfo.FileName =
+                    @"..\Int5.DnD3D.WebClient\Int5.DnD3D.WebClient\bin\Debug\net6.0\Int5.DnD3D.WebClient.exe";
+                DataContainer.WebserviceConnection.Start();
+                
+                while (!File.Exists(Path.GetTempPath() + @$"/DnD/0") && DataContainer.UserNameValid)
+                {
+                }
+
+                Debug.Log(DataContainer.UserNameValid);
+                if (DataContainer.UserNameValid)
+                {
+                    string content = File.ReadAllText(Path.GetTempPath() + "/DnD/" + 0);
+                    DataContainer.ClientId = Guid.Parse(content.Substring(1, content.Length - 1));
+                    File.Delete(Path.GetTempPath() + "/DnD/0");
+                    SceneManager.LoadScene("SampleScene");
+                }
             }
             else
             {
@@ -55,6 +66,7 @@ public class Login : MonoBehaviour
             popup.ShowPopup("Server not found. Please check if the server is running and the ip is correct.");
         }
     }
+
     private string Connect(WebRequest webRequest)
     {
         using var response = webRequest.GetResponse();
